Reorganise code placing flickity inside of displayCategories.js and placing isotope inside of isotope.js
