+++ src/javascript/categoriesController.js
@@ -3,7 +3,7 @@ import { formatCategoryTitle } from './formatText.js';
 import { flashError } from './flashError.js';
 import { toggleLoadingScreen } from './startScreen.js';
 import displayCategories from './displayCategories.js';
-import { iso } from '../index.js';
+import iso from './isotope.js';
 import { toggleAddCategoryMenu } from './toggleForms.js';
 import { toggleDeleteCategoryMenu } from './toggleMenus.js';
