+++ src/javascript/displayTodos.js
@@ -1,4 +1,4 @@
-import { iso } from '../index.js';
+import iso from './isotope.js';
 import { getTodosArray, getCategoriesArray } from './database.js';
 import { setTodoMenu, setDeleteTodoMenu } from './toggleMenus.js';
 import { setEditTodoMenu } from './toggleForms.js';
