+++ src/javascript/startApp.js
@@ -8,7 +8,7 @@ import setShowAllCategoriesBtn from './showAllCategories.js';
 import { setCategoriesController } from './categoriesController.js';
 import { setTodosController } from './todosController.js';
 import { displayAllTodos } from './displayTodos.js';
-import { iso } from '../index.js';
+import iso from './isotope.js';
 
 const start = () => {
   launchStartScreen();
