+++ src/javascript/todosController.js
@@ -5,7 +5,7 @@ import {
 } from './database.js';
 import { toggleAddTodoMenu, toggleEditTodoMenu } from './toggleForms.js';
 import { toggleDeleteTodoMenu, toggleDeleteAllTodoMenu } from './toggleMenus.js';
-import { iso } from '../index.js';
+import iso from './isotope.js';
 import { formatTodoTitle, formatTodoDescription } from './formatText.js';
 import { flashError } from './flashError.js';
